Move register_ap_actor_rest_field to Post_Types class
The register_ap_actor_rest_field method and its related rest_api_init hook have been moved from the Activitypub class to the Post_Types class. This change centralizes REST field registration logic related to post types, improving code organization.
